claims-automation-advanced/requirements.txt
pytesseract
Pillow
pdfplumber